+++ deprecated/mcp_sampling_optimized.py
