+++ deprecated/mcp_sampling_patch.py
