+++ tests/security/SECURITY_REGRESSION_TEST_SUMMARY.md
@@ -1,0 +1,174 @@
+# Security Regression Test Summary
+
+## Overview
+Comprehensive security regression test suite for PR #190 vulnerabilities.
+
+**Test File**: `tests/security/test_security_regression.py`
+
+## Test Coverage
+
+### CRITICAL Vulnerabilities (5 tests)
+
+1. **test_code_injection_via_workspace_data** ✅
+   - Tests prevention of code injection via unsanitized workspace data
+   - Validates that prompt injection attempts are sanitized
+   - Ensures malicious comments/instructions are redacted
+
+2. **test_template_injection_prevention** ✅
+   - Tests template injection vulnerability prevention
+   - Validates safe template rendering without code execution
+   - Ensures expression evaluation is blocked
+
+3. **test_secrets_exposure_prevention** ✅
+   - Tests that API keys and secrets are redacted
+   - Validates multiple secret patterns (AWS, GitHub, OpenAI, etc.)
+   - Ensures private keys are not exposed
+
+4. **test_unrestricted_file_access_prevention** ✅
+   - Tests file system access restrictions
+   - Validates workspace isolation
+   - Ensures sensitive system files cannot be accessed
+
+5. **test_rate_limiting_enforcement** ✅
+   - Tests rate limiting on API endpoints
+   - Validates burst protection
+   - Ensures DoS prevention
+
+### HIGH Vulnerabilities (3 tests)
+
+6. **test_path_traversal_prevention** ✅
+   - Tests path traversal attack prevention
+   - Validates various encoding attempts are blocked
+   - Ensures directory traversal is impossible
+
+7. **test_redos_prevention** ✅
+   - Tests ReDoS (Regex Denial of Service) prevention
+   - Validates catastrophic backtracking protection
+   - Ensures regex operations complete in reasonable time
+
+8. **test_input_validation_comprehensive** ✅
+   - Tests comprehensive input validation
+   - Validates null bytes, control characters handling
+   - Ensures malicious inputs are sanitized
+
+### MEDIUM Vulnerabilities (4 tests)
+
+9. **test_insecure_randomness_mitigation** ✅
+   - Tests secure random number generation
+   - Validates cache key entropy
+   - Ensures unpredictable session identifiers
+
+10. **test_verbose_error_messages_sanitization** ✅
+    - Tests error message sanitization
+    - Validates no sensitive information leakage
+    - Ensures generic error responses
+
+11. **test_authentication_requirement** ✅
+    - Tests authentication tracking implementation
+    - Validates user_id parameter presence
+    - Ensures per-user rate limiting
+
+12. **test_cache_poisoning_prevention** ✅
+    - Tests cache key generation security
+    - Validates injection prevention in cache keys
+    - Ensures deterministic but secure caching
+
+### Integration Tests
+
+13. **test_complete_security_pipeline** ✅
+    - End-to-end security pipeline validation
+    - Tests all mitigations working together
+    - Validates complex attack scenarios
+
+14. **test_security_patches_applied_on_import** ✅
+    - Tests automatic patch application
+    - Validates runtime patching mechanism
+    - Ensures patches don't break functionality
+
+15. **test_security_monitoring_and_logging** ✅
+    - Tests security event logging
+    - Validates audit trail creation
+    - Ensures monitoring capabilities
+
+### Additional Tests with Fixtures
+
+16. **test_workspace_isolation** ✅
+    - Tests workspace directory isolation
+    - Validates file access boundaries
+    - Uses mock secure environment fixture
+
+17. **test_symlink_attack_prevention** ✅
+    - Tests symlink attack prevention
+    - Validates symlinks cannot escape workspace
+    - Ensures file reference security
+
+## Test Execution
+
+### Run All Security Regression Tests
+```bash
+pytest tests/security/test_security_regression.py -v
+```
+
+### Run by Severity
+```bash
+# Critical vulnerabilities only
+pytest tests/security/test_security_regression.py -m critical -v
+
+# High severity only  
+pytest tests/security/test_security_regression.py -m high -v
+
+# Medium severity only
+pytest tests/security/test_security_regression.py -m medium -v
+```
+
+### Run Integration Tests
+```bash
+pytest tests/security/test_security_regression.py -m integration -v
+```
+
+## Expected Behavior
+
+### Without Security Patches
+- Tests should FAIL, demonstrating vulnerabilities exist
+- This validates the tests correctly identify security issues
+
+### With Security Patches Applied
+- All tests should PASS
+- This confirms vulnerabilities are properly mitigated
+- Runtime patching mechanism should activate automatically
+
+## Test Design Principles
+
+1. **Positive & Negative Testing**: Each test includes both:
+   - Attack attempts that should be blocked
+   - Legitimate operations that should succeed
+
+2. **Realistic Attack Vectors**: Tests use actual attack patterns:
+   - Real injection payloads
+   - Common secret formats
+   - Known traversal techniques
+
+3. **Comprehensive Coverage**: Tests cover:
+   - Direct attacks
+   - Encoded/obfuscated attacks
+   - Combined attack scenarios
+
+4. **Clear Documentation**: Each test includes:
+   - Vulnerability description
+   - Line number reference from PR #190
+   - Expected mitigation behavior
+
+## Current Status
+
+- **Test File Created**: ✅ Complete
+- **All 12 Vulnerabilities Covered**: ✅ Yes
+- **Integration Tests**: ✅ Included
+- **Fixture-based Tests**: ✅ Included
+- **Documentation**: ✅ Complete
+
+## Notes
+
+- Tests use pytest marks for categorization (security, regression, critical, high, medium, integration)
+- Async tests properly marked with @pytest.mark.asyncio
+- Tests designed to work with both patched and unpatched code
+- Comprehensive assertions validate security controls are working
